+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -15,6 +15,7 @@
 - TITLE instruction name field support
 - Produce only a warning when immediate operand overflows in machine instructions
 - Relax PROCESS operand validation
+- Syntax error while parsing the END language operands
 
 ## [1.4.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.3.0...1.4.0) (2022-08-30)